+++ scripts/main.js
@@ -195,7 +195,7 @@ function createWindow () {
 
 function accelerators_setup () {
   /* istanbul ignore next */
-  globalShortcut.register('Alt+Space', () => { mainWindow.webContents.send('svg_reset_zoom') })
+  globalShortcut.register('Alt+Home', () => { mainWindow.webContents.send('svg_reset_zoom') })
   /* istanbul ignore next */
   globalShortcut.register('Alt+0', () => { mainWindow.webContents.send('svg_reset_zoom') })
   /* istanbul ignore next */
@@ -235,6 +235,7 @@ app.on('ready', () => {
       newVer: { type: 'boolean', alias: 'V', desc: 'Check for new release', default: undefined }, // has a setting
       select: { type: 'string', alias: 's', desc: 'Selection criteria', default: undefined },
       idOnly: { type: 'boolean', alias: 'i', desc: 'Search id only', default: false },
+      limitDepth: { type: 'boolean', alias: '1', desc: 'Limit reachable nodes to 1 level', default: false }, //rq: ->(rq_limited_walk_cl)
       exclIds: { type: 'string', alias: 'e', desc: 'Excluded ids, comma separated', default: undefined },
       inclRejected: { type: 'boolean', alias: 'R', desc: 'Include rejected specobjects', default: false },
       exclDoctypes: { type: 'string', alias: 'T', desc: 'Excluded doctypes, comma separated', default: undefined },